separated useCrossFusePools hook
